alteração nas classes de controller, funcionarios,repository,ClienteUserCase
