+++ src/main/java/com/mypet/mypet/repositories/FuncionarioRepository.java
@@ -1,0 +1,11 @@
+package com.mypet.mypet.repositories;
+
+import com.mypet.mypet.domain.model.Funcionarios;
+import com.mypet.mypet.domain.model.Pessoas;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+public interface FuncionarioRepository extends JpaRepository<Funcionarios,Long> {
+
+}
