+++ src/main/java/com/mypet/mypet/userCase/ClienteUserCase.java
@@ -1,8 +1,9 @@
 package com.mypet.mypet.userCase;
 
+import com.mypet.mypet.domain.model.Funcionarios;
 import com.mypet.mypet.domain.model.Pessoas;
 import com.mypet.mypet.domain.dto.ClienteDTO;
-import com.mypet.mypet.repositories.ClienteRepository;
+import com.mypet.mypet.repositories.FuncionarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,9 +12,9 @@ import java.util.List;
 @Service
 public class ClienteUserCase {
     @Autowired
-    private ClienteRepository clienteRepository;
+    private FuncionarioRepository funcionarioRepository;
     public List<ClienteDTO> findAll(){
-    List<Pessoas> result = clienteRepository.findAll();
+    List<Funcionarios> result = funcionarioRepository.findAll();
     List<ClienteDTO> dto = result.stream().map(x -> new ClienteDTO(x)).toList();
     return dto;
 }
